+++ aiohttp_sqlalchemy/__init__.py
@@ -46,7 +46,7 @@ from aiohttp_sqlalchemy.views import (
     SAView,
 )
 
-__version__ = '0.25.0'
+__version__ = '0.26.0'
 
 __all__ = [
     'DEFAULT_KEY',